requirements.txt
fastapi==0.95.0
uvicorn==0.18.2
psycopg2-binary==2.9.3
pydantic==1.10.5
python-dotenv==0.21.0
sqlalchemy==2.0.10
pyjwt==2.6.0
python-multipart==0.0.5
bcrypt==4.0.1
pytest>=7.0.0